:truck: Rename config files
