+++ src/modules/game/config/renderer.ts
